+++ src/Build5Nines.SharpVector/VectorCompare/CosineSimilarityVectorComparerAsync.cs
@@ -1,4 +1,5 @@
 using System.Reflection;
+using System.Runtime.InteropServices;
 
 namespace Build5Nines.SharpVector.VectorCompare;
 
@@ -56,22 +57,12 @@ public class CosineSimilarityVectorComparer : IVectorComparer
         return dotProduct / (magnitudeA * magnitudeB);
     }
 
-    public IEnumerable<VectorTextResultItem<TDocument, TMetadata>> Sort<TDocument, TMetadata>(IEnumerable<VectorTextResultItem<TDocument, TMetadata>> results)
+    public IEnumerable<IVectorTextResultItem<TId, TDocument, TMetadata>> Sort<TId, TDocument, TMetadata>(IEnumerable<IVectorTextResultItem<TId, TDocument, TMetadata>> results)
     {
         return results.OrderByDescending(s => s.VectorComparison);
     }
 
-    public IEnumerable<VectorTextResultItem<TMetadata>> Sort<TMetadata>(IEnumerable<VectorTextResultItem<TMetadata>> results)
-    {
-        return results.OrderByDescending(s => s.VectorComparison);
-    }
-
-    public async Task<IEnumerable<VectorTextResultItem<TDocument, TMetadata>>> SortAsync<TDocument, TMetadata>(IEnumerable<VectorTextResultItem<TDocument, TMetadata>> results)
-    {
-        return await Task.Run(() => Sort(results));
-    }
-
-    public async Task<IEnumerable<VectorTextResultItem<TMetadata>>> SortAsync<TMetadata>(IEnumerable<VectorTextResultItem<TMetadata>> results)
+    public async Task<IEnumerable<IVectorTextResultItem<TId, TDocument, TMetadata>>> SortAsync<TId, TDocument, TMetadata>(IEnumerable<IVectorTextResultItem<TId, TDocument, TMetadata>> results)
     {
         return await Task.Run(() => Sort(results));
     }