+++ src/Build5Nines.SharpVector/VectorTextResultItem.cs
@@ -1,4 +1,5 @@
 using System.Collections.Immutable;
+using Build5Nines.SharpVector.Id;
 
 namespace Build5Nines.SharpVector;
 
@@ -10,21 +11,32 @@ public interface IVectorTextResultItem<TDocument, TMetadata>
     float VectorComparison { get; }
 }
 
+public interface IVectorTextResultItem<TId, TDocument, TMetadata>
+    : IVectorTextResultItem<TDocument, TMetadata>
+{
+    TId Id { get; }
+}
+
 public interface IVectorTextResultItem<TMetadata>
- : IVectorTextResultItem<string, TMetadata>
+ : IVectorTextResultItem<string, TMetadata>, IVectorTextResultItem<int, string, TMetadata>
 { }
 
-public class VectorTextResultItem<TDocument, TMetadata> : IVectorTextResultItem<TDocument, TMetadata>
+public class VectorTextResultItem<TId, TDocument, TMetadata>
+    : IVectorTextResultItem<TDocument, TMetadata>, IVectorTextResultItem<TId, TDocument, TMetadata>
 {
     private IVectorTextItem<TDocument, TMetadata> _item;
-    public VectorTextResultItem(IVectorTextItem<TDocument, TMetadata> item, float vectorComparison)
+    private TId _id;
+
+    public VectorTextResultItem(TId id, IVectorTextItem<TDocument, TMetadata> item, float vectorComparison)
     {
+        _id = id;
         _item = item;
         VectorComparison = vectorComparison;
     }
     
     public TDocument Text { get => _item.Text; }
     public TMetadata? Metadata { get => _item.Metadata; }
+    public TId Id { get => _id; }
 
     public ImmutableArray<float> Vectors { get => ImmutableArray.Create(_item.Vector); }
 
@@ -32,9 +44,10 @@ public class VectorTextResultItem<TDocument, TMetadata> : IVectorTextResultItem<
 }
 
 public class VectorTextResultItem<TMetadata>
- : VectorTextResultItem<string, TMetadata>, IVectorTextResultItem<TMetadata>
+ : VectorTextResultItem<int, string, TMetadata>, IVectorTextResultItem<TMetadata>
 {
-    public VectorTextResultItem(IVectorTextItem<string, TMetadata> item, float vectorComparison)
-        : base(item, vectorComparison)
+    public VectorTextResultItem(int id, IVectorTextItem<string, TMetadata> item, float vectorComparison)
+        : base(id, item, vectorComparison)
     { }
 }
+