+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v2.1.0
+
+Add:
+
+- Added `VectorTextResultItem.Id` property so it's easy to get the database ID for search results if necessary.
+- `IVectorDatabase` now inherits from `IEnumerable` so you can easily look through the texts documents that have been added to the database.
+
+Fixed:
+
+- Fixed text tokenization to correctly remove special characters
+- Update `BasicTextPreprocessor` to support Emoji characters too
+- Refactorings for more Clean Code
+
+Breaking Changes:
+
+- The `.Search` and `.SearchAsync` methods now return a `IVectorTextResultItem<TId, TDocument, TMetadata>` instead of `VectorTextResultItem<TDocument, TMetadata>`. If you're using things like the documentation shows, then you wont see any changes or have any issues with this update.
+
 ## v2.0.4 (2025-04-16)
 
 Add: